新增OrthoSmallOrder & NoResult

+++ OrderManagerNew/AirDental_UserControls/AirD_orthoBase.xaml.cs
@@ -23,6 +23,9 @@ namespace OrderManagerNew.AirDental_UserControls
         //委派到MainWindow.xaml.cs裡面CaseHandler_Ortho_showSingleProject()
         public delegate void AirD_orthoBaseEventHandler(int projectIndex);
         public event AirD_orthoBaseEventHandler SetAirDentalProjectShow;
+        //委派到MainWindow.xaml.cs裡面的CaseHandler_Ortho_showDetail()
+        public delegate void AirD_orthoBaseEventHandler2(int BaseCaseIndex, int SmallCaseIndex);
+        public event AirD_orthoBaseEventHandler2 SetSmallOrderDetailShow;
 
         public Dll_Airdental.Main orthoBase_AirDental;
         public List<Dll_Airdental.Main._orthoOrder> Orderlist_Ortho;
@@ -30,7 +33,7 @@ namespace OrderManagerNew.AirDental_UserControls
         AirD_orthoProject orthoProjectInfo;
         bool IsFocusCase = false;   //smallCase目前是否為攤開狀態
 
-        public class AirD_orthoOrder
+        /*public class AirD_orthoOrder
         {
             //內部專案Oid
             public string Oid { get; set; }
@@ -59,7 +62,7 @@ namespace OrderManagerNew.AirDental_UserControls
                 CreateDate = new DateTimeOffset();
                 Viewerurl = "";
             }
-        }
+        }*/
         public class AirD_orthoProject
         {
             //外部專案Pid
@@ -78,7 +81,7 @@ namespace OrderManagerNew.AirDental_UserControls
             public string PatientAvatar { get; set; }
             public string TxTreatedArch { get; set; }
             public string ProductType { get; set; }
-            public AirD_orthoOrder[] List_orthoOrder { get; set; }
+            public List<AirD_orthoSmallOrder> List_orthoOrder { get; set; }
 
             public AirD_orthoProject()
             {
@@ -124,7 +127,23 @@ namespace OrderManagerNew.AirDental_UserControls
 
         private void LoadOrthoOrders()
         {
-            //TODO明天繼續
+            orthoProjectInfo.List_orthoOrder = new List<AirD_orthoSmallOrder>();
+            if (Properties.Settings.Default.showCloudOrderNumbers < 1)
+                Properties.Settings.Default.showCloudOrderNumbers = 5;
+
+            int totalCount = -1;
+            if (Orderlist_Ortho.Count < Properties.Settings.Default.showCloudOrderNumbers)
+                totalCount = Orderlist_Ortho.Count;
+            else
+                totalCount = Properties.Settings.Default.showCloudOrderNumbers;
+
+            for(int i=0; i<totalCount; i++)
+            {
+                AirDental_UserControls.AirD_orthoSmallOrder TmpOrthoSmallOrder = new AirD_orthoSmallOrder();
+                TmpOrthoSmallOrder.SetOrderCaseShow += new AirD_orthoSmallOrder.orthoOrderEventHandler(SmallOrderHandler);
+                TmpOrthoSmallOrder.SetOrderInfo(Orderlist_Ortho[i], i);
+                orthoProjectInfo.List_orthoOrder.Add(TmpOrthoSmallOrder);
+            }
         }
 
         public void SetProjectInfo(Dll_Airdental.Main._orthoProject Import, int Index)
@@ -150,7 +169,7 @@ namespace OrderManagerNew.AirDental_UserControls
             orthoProjectIndex = Index;
 
             label_orderID.Content = orthoProjectInfo.SerialNumber;
-            label_designStep.Content = orthoProjectInfo.Stage_String;
+            label_designStep.Content = TranslationSource.Instance[orthoProjectInfo.Group] + orthoProjectInfo.Action_String + orthoProjectInfo.Stage_String;
             label_patientName.Content = orthoProjectInfo.Patient;
             label_modifyDate.Content = orthoProjectInfo.ModifyDate.DateTime.ToLongDateString() + orthoProjectInfo.ModifyDate.DateTime.ToLongTimeString();
 
@@ -171,6 +190,22 @@ namespace OrderManagerNew.AirDental_UserControls
         }
 
         /// <summary>
+        /// 使用者點擊SmallCase時事件
+        /// </summary>
+        /// <param name="SmallcaseIndex">SmallCase的Index</param>
+        private void SmallOrderHandler(int SmallorderIndex)
+        {
+            SetSmallOrderDetailShow(orthoProjectIndex, SmallorderIndex);  //MainWindow顯示Small Case Detail
+            for (int i = 0; i < orthoProjectInfo.List_orthoOrder.Count; i++)
+            {
+                if (i == SmallorderIndex)
+                    continue;
+
+                orthoProjectInfo.List_orthoOrder[i].SetCaseFocusStatus(false);
+            }
+        }
+
+        /// <summary>
         /// 設定Case的Focus狀態
         /// </summary>
         /// <param name="isFocused">是否要Focus</param>
@@ -184,10 +219,10 @@ namespace OrderManagerNew.AirDental_UserControls
                         //執行攤開
                         if (orthoProjectInfo.List_orthoOrder != null)
                         {
-                            /*foreach (Order_orthoSmallcase OrthoCase in orthoInfo.List_smallcase)
+                            foreach (AirD_orthoSmallOrder OrthoOrder in orthoProjectInfo.List_orthoOrder)
                             {
-                                stackpanel_Ortho.Children.Add(OrthoCase);
-                            }*/
+                                stackpanel_Ortho.Children.Add(OrthoOrder);
+                            }
                         }
                         else
                         {
@@ -195,13 +230,13 @@ namespace OrderManagerNew.AirDental_UserControls
                             Mouse.OverrideCursor = Cursors.Wait;
                             GetOrthoOrder();
                             Mouse.OverrideCursor = Cursors.Arrow;
-                            /*if (orthoInfo.List_smallcase.Count > 0)
+                            if (orthoProjectInfo.List_orthoOrder != null)
                             {
-                                foreach (Order_orthoSmallcase OrthoCase in orthoInfo.List_smallcase)
+                                foreach (AirD_orthoSmallOrder OrthoOrder in orthoProjectInfo.List_orthoOrder)
                                 {
-                                    stackpanel_Ortho.Children.Add(OrthoCase);
+                                    stackpanel_Ortho.Children.Add(OrthoOrder);
                                 }
-                            }*/
+                            }
                         }
                         IsFocusCase = true;
                         break;
@@ -210,19 +245,25 @@ namespace OrderManagerNew.AirDental_UserControls
                     {
                         background_orthoBase.Fill = Brushes.White;
                         //收回
-                        /*if (orthoProjectInfo.List_orthoOrder.Length > 0)
+                        if (orthoProjectInfo.List_orthoOrder != null)
                         {
                             for (int i = 1; i < stackpanel_Ortho.Children.Count; i++)
                             {
-                                ((Order_orthoSmallcase)stackpanel_Ortho.Children[i]).SetCaseFocusStatus(false);
+                                ((AirD_orthoSmallOrder)stackpanel_Ortho.Children[i]).SetCaseFocusStatus(false);
                             }
 
                             stackpanel_Ortho.Children.RemoveRange(1, (stackpanel_Ortho.Children.Count - 1));
-                        }*/
+                        }
                         IsFocusCase = false;
                         break;
                     }
             }
+        }
+        
+
+        private void Click_AirdentalWeb(object sender, RoutedEventArgs e)
+        {
+
         }
 
         private void PMDown_StackPanelMain(object sender, MouseButtonEventArgs e)
@@ -243,11 +284,6 @@ namespace OrderManagerNew.AirDental_UserControls
                     SetCaseFocusStatus(false);
                 }
             }
-        }
-
-        private void Click_AirdentalWeb(object sender, RoutedEventArgs e)
-        {
-
         }
     }
 }
